Apply sofft delete for category function

+++ src/main/java/com/ecommerce/backend/service/CategoryService.java
@@ -3,6 +3,7 @@ package com.ecommerce.backend.service;
 import com.ecommerce.backend.entity.Category;
 import com.ecommerce.backend.exception.ResourceNotFoundException;
 import com.ecommerce.backend.model.request.CategoryRequest;
+import com.ecommerce.backend.model.request.SoftDeleteRequest;
 import com.ecommerce.backend.model.response.CategoryResponse;
 import com.ecommerce.backend.repository.CategoryRepository;
 import jakarta.persistence.EntityManager;
@@ -47,16 +48,17 @@ public class CategoryService {
         return new CategoryResponse(savedCategory.getName(), savedCategory.getDeleted());
     }
 
-    public void deleteById(UUID categoryId) {
-        categoryRepository.deleteById(categoryId);
+    public void deleteById(UUID categoryId, SoftDeleteRequest request) {
+        if (request.isSoftDelete()) {
+            categoryRepository.softDeleteById(categoryId);
+        }
+        else {
+            categoryRepository.deleteById(categoryId);
+        }
     }
 
     public List<CategoryResponse> getAllForUser() {
-        Session session = entityManager.unwrap(Session.class);
-        Filter filter = session.enableFilter("deletedCategory");
-        filter.setParameter("isDeleted", Boolean.FALSE);
-        List<Category> categories = categoryRepository.findAll();
-        session.disableFilter("deletedCategory");
+        List<Category> categories = categoryRepository.findByDeletedFalse();
         return categories
                 .stream()
                 .map(category -> new CategoryResponse(category.getName(), category.getDeleted()))
